pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "breadboard"
version = "0.1.0"
description = "Real-time stock pricing dashboard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.28.0",
    "clickhouse-driver>=0.2.6",
    "apscheduler>=3.10.4",
    "websockets>=12.0",
    "yfinance>=0.2.32",
    "protobuf>=4.25.0",
    "pandas>=2.1.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[tool.setuptools]
packages = ["backend", "dashboard"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
